Move the embedder interface into lightly-studio-embed
`lightly-studio` and `lightly-studio-embed` each declared their own embedder
classes. The two sets had to stay the same by hand. This change moves the shared
classes into the workspace member and imports them back. There is now one
definition, so a model written against either package has the same type.

`lightly_studio/embed/embedder.py` and `lightly_studio/embed/types.py` move
without a shim. `types.py` does not change at all. `embedder.py` changes one
import line. No class, member, method or docstring is added, removed or
reordered.

All 25 callers import from `lightly_studio_embed`. `lightly_studio` still
exports `EmbeddingSpaceSpec` and `ImageCrop`, and now also `EmbeddingResult`.
The documented examples therefore stay in one namespace and do not name the
second package.

The docs handler needs the source of the member on its path. Without it griffe
cannot load a moved class, and three sections of `api/embeddings.md` render as
an empty module heading. Those sections now point at the module that defines
the types.

The package gains a numpy dependency, and the README and the pyproject comment
are corrected. `EmbeddingResult.embeddings` must stay an array, because the
ingest paths of LightlyStudio index and slice it. A nested sequence cannot do
that. Torch, CUDA and `lightly-studio` stay out of the dependencies.
