+++ lightly_studio/src/lightly_studio/__init__.py
@@ -26,7 +26,7 @@ from lightly_studio.dataset.embedding_generator import (
     ImageEmbeddingGenerator,
     VideoEmbeddingGenerator,
 )
-from lightly_studio.embed.types import EmbeddingSpaceSpec, ImageCrop
+from lightly_studio_embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
 from lightly_studio.dataset.embedding_manager import set_default_embedding_model
 from lightly_studio.models.collection import SampleType
 from lightly_studio.enterprise import connect
@@ -42,6 +42,7 @@ __all__ = [
     "CreateMcap",
     "CreateVideo",
     "EmbeddingGenerator",
+    "EmbeddingResult",
     "EmbeddingSpaceSpec",
     "GroupDataset",
     "ImageCrop",
