+++ lightly_studio/src/lightly_studio/dataset/perception_encoder_embedding_generator.py
@@ -8,6 +8,7 @@ import fsspec
 import numpy as np
 import torch
 from av import FFmpegError, container
+from lightly_studio_embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
 from numpy.typing import NDArray
 from PIL import Image
 from tqdm import tqdm
@@ -21,7 +22,6 @@ from lightly_studio.core.file_outcome_report import (
 from lightly_studio.dataset.env import LIGHTLY_STUDIO_MODEL_CACHE_DIR
 from lightly_studio.embed import image_crop_embedding, image_embedding
 from lightly_studio.embed.image_embedding import EmbeddingContext
-from lightly_studio.embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
 from lightly_studio.utils import batching
 from lightly_studio.vendor.perception_encoder.vision_encoder import pe, transforms
