+++ lightly_studio/src/lightly_studio/embed/image_crop_embedding.py
@@ -8,6 +8,7 @@ from typing import Any
 import fsspec
 import numpy as np
 import torch
+from lightly_studio_embed.types import EmbeddingResult, ImageCrop
 from numpy.typing import NDArray
 from PIL import Image
 from tqdm import tqdm
@@ -18,7 +19,6 @@ from lightly_studio.core.file_outcome_report import (
     FileOutcomeReport,
 )
 from lightly_studio.embed.image_embedding import EmbeddingContext
-from lightly_studio.embed.types import EmbeddingResult, ImageCrop
 from lightly_studio.utils import executor, parallelize
