+++ lightly_studio/src/lightly_studio/embed/random_embedder.py
@@ -8,9 +8,7 @@ results are random.
 from __future__ import annotations
 
 import numpy as np
-from PIL.Image import Image
-
-from lightly_studio.embed.embedder import (
+from lightly_studio_embed.embedder import (
     ImageBytesEmbedder,
     ImageCropPathEmbedder,
     ImagePathEmbedder,
@@ -18,7 +16,8 @@ from lightly_studio.embed.embedder import (
     TextEmbedder,
     VideoPathEmbedder,
 )
-from lightly_studio.embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
+from lightly_studio_embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
+from PIL.Image import Image
 
 
 class RandomEmbedder(
