+++ lightly_studio/src/lightly_studio/examples/example_custom_embedding_model.py
@@ -16,6 +16,7 @@ from pathlib import Path
 import numpy as np
 import torch
 from environs import Env
+from lightly_studio_embed.types import EmbeddingResult
 from numpy.typing import NDArray
 from PIL import Image
 
@@ -25,7 +26,6 @@ from lightly_studio.dataset import file_utils
 from lightly_studio.dataset.env import LIGHTLY_STUDIO_MODEL_CACHE_DIR
 from lightly_studio.embed import image_crop_embedding, image_embedding
 from lightly_studio.embed.image_embedding import EmbeddingContext
-from lightly_studio.embed.types import EmbeddingResult
 from lightly_studio.vendor import mobileclip
 
 MODEL_NAME = "mobileclip_s0"
