+++ lightly_studio/src/lightly_studio/examples/example_load_existing_embeddings.py
@@ -14,12 +14,12 @@ from pathlib import Path
 
 import numpy as np
 from environs import Env
+from lightly_studio_embed.types import EmbeddingResult
 from numpy.typing import NDArray
 from PIL import Image
 
 import lightly_studio as ls
 from lightly_studio.database import db_manager
-from lightly_studio.embed.types import EmbeddingResult
 
 IMAGE_SUFFIXES = {".jpg", ".jpeg", ".png", ".bmp", ".webp"}
 EMBEDDING_DIMENSION = 64
