+++ lightly_studio/src/lightly_studio/resolvers/annotation_resolver/get_annotation_crops.py
@@ -6,9 +6,9 @@ import logging
 from dataclasses import dataclass
 from uuid import UUID
 
+from lightly_studio_embed.types import ImageCrop
 from sqlmodel import Session, col, select
 
-from lightly_studio.embed.types import ImageCrop
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable
 from lightly_studio.models.annotation.object_detection import ObjectDetectionAnnotationTable
 from lightly_studio.models.annotation.segmentation import SegmentationAnnotationTable
