+++ lightly_studio/tests/dataset/test_embedding_generator.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from lightly_studio_embed.types import ImageCrop
+
 from lightly_studio.dataset.embedding_generator import RandomEmbeddingGenerator
-from lightly_studio.embed.types import ImageCrop
 
 
 class TestRandomEmbeddingGeneratorCrops:
