+++ lightly_studio/tests/embed/test_image_crop_embedding.py
@@ -9,6 +9,7 @@ import fsspec
 import numpy as np
 import pytest
 import torch
+from lightly_studio_embed.types import ImageCrop
 from numpy.typing import NDArray
 from PIL import Image
 from pytest_mock import MockerFixture
@@ -16,7 +17,6 @@ from pytest_mock import MockerFixture
 from lightly_studio.core.file_outcome_report import AllInputFilesFailedError
 from lightly_studio.embed import image_crop_embedding
 from lightly_studio.embed.image_embedding import EmbeddingContext
-from lightly_studio.embed.types import ImageCrop
 from lightly_studio.utils import executor
