+++ lightly_studio/tests/embed/test_mobileclip_embedder.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import numpy as np
 import torch
+from lightly_studio_embed.types import ImageCrop
 from PIL import Image
 
 from lightly_studio.embed.mobileclip_embedder import (
@@ -11,7 +12,6 @@ from lightly_studio.embed.mobileclip_embedder import (
     MODEL_NAME,
     MobileCLIPEmbedder,
 )
-from lightly_studio.embed.types import ImageCrop
 
 FIXTURES_DIR = Path(__file__).parent.parent / "fixtures"
