+++ lightly_studio_embed/README.md
@@ -3,11 +3,14 @@
 Serve your own embedding model to [LightlyStudio](https://github.com/lightly-ai/lightly-studio)
 over HTTP, so that your model's weights never leave your machine.
 
-The package deliberately depends on nothing but an HTTP server, so that it installs next to your
-own CUDA and torch pins.
+The package depends on an HTTP server, numpy and Pillow. It does not depend on torch, CUDA or
+LightlyStudio. You can therefore install it next to your own pins.
 
-There is no public API yet, and nothing is published to PyPI: this only reserves the package. Once
-it is released, installing it will be:
+`lightly_studio_embed.embedder` holds the base classes for a model. `lightly_studio_embed.types`
+holds the values that these classes receive and return. LightlyStudio uses the same classes for
+its own embedders.
+
+Nothing is published to PyPI yet. Once it is released, installing it will be:
 
 ```bash
 pip install lightly-studio-embed
