+++ lightly_studio_embed/src/lightly_studio_embed/__init__.py
@@ -1,1 +1,27 @@
 """Serve your own embedding model to LightlyStudio over HTTP."""
+
+from lightly_studio_embed.embedder import (
+    Capability,
+    Embedder,
+    ImageBytesEmbedder,
+    ImageCropPathEmbedder,
+    ImagePathEmbedder,
+    ImagePILEmbedder,
+    TextEmbedder,
+    VideoPathEmbedder,
+)
+from lightly_studio_embed.types import EmbeddingResult, EmbeddingSpaceSpec, ImageCrop
+
+__all__ = [
+    "Capability",
+    "Embedder",
+    "EmbeddingResult",
+    "EmbeddingSpaceSpec",
+    "ImageBytesEmbedder",
+    "ImageCrop",
+    "ImageCropPathEmbedder",
+    "ImagePILEmbedder",
+    "ImagePathEmbedder",
+    "TextEmbedder",
+    "VideoPathEmbedder",
+]
