+++ lightly_studio_embed/src/lightly_studio_embed/types.py
